fix paths & rename numpad
